# ecommerce/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Vendor AI E-commerce API
  version: 1.0.0
  description: REST API for the multi-vendor AI e-commerce platform.
paths:
...
  /api/v1/admin/brands/:
    get:
      operationId: admin_brands_list
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandListResponse'
          description: ''
    post:
      operationId: admin_brands_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BrandCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrandCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
          description: ''
  /api/v1/admin/brands/{brand_id}/:
    get:
      operationId: admin_brands_retrieve
      parameters:
      - in: path
        name: brand_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
          description: ''
    patch:
      operationId: admin_brands_update
      parameters:
      - in: path
        name: brand_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedBrandUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedBrandUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedBrandUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
          description: ''
    delete:
      operationId: admin_brands_delete
      parameters:
      - in: path
        name: brand_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
          description: ''
  /api/v1/admin/categories/:
    get:
      operationId: admin_categories_list
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
          description: ''
    post:
      operationId: admin_categories_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
          description: ''
  /api/v1/admin/categories/{category_id}/:
    get:
      operationId: admin_categories_retrieve
      parameters:
      - in: path
        name: category_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
          description: ''
    patch:
      operationId: admin_categories_update
      parameters:
      - in: path
        name: category_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCategoryUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCategoryUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCategoryUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
          description: ''
    delete:
      operationId: admin_categories_delete
      parameters:
      - in: path
        name: category_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
          description: ''
  /api/v1/admin/categories/reorder/:
    post:
      operationId: admin_categories_reorder
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryReorder'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CategoryReorder'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryReorder'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryReorderResponse'
          description: ''
...
  /api/v1/brands/:
    get:
      operationId: public_brands_list
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandListResponse'
          description: ''
  /api/v1/categories/:
    get:
      operationId: public_categories_tree
      tags:
      - v1
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryTreeResponse'
          description: ''
...
components:
  schemas:
...
    BrandCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        slug:
          type: string
          maxLength: 180
          pattern: ^[-a-zA-Z0-9_]+$
        logo_url:
          nullable: true
          oneOf:
          - type: string
            format: uri
          - type: string
            maxLength: 0
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
      required:
      - name
    BrandList:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        logo_url:
          type: string
          readOnly: true
          nullable: true
        description:
          type: string
          readOnly: true
          nullable: true
        is_active:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - description
      - id
      - is_active
      - logo_url
      - name
      - slug
      - updated_at
    BrandListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/BrandList'
        meta:
          $ref: '#/components/schemas/CatalogPaginationMeta'
      required:
      - data
      - message
      - meta
      - success
    BrandResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/BrandList'
      required:
      - data
      - message
      - success
...
    CatalogPaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
      required:
      - page
      - page_size
      - total_items
      - total_pages
    CategoryCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        slug:
          type: string
          maxLength: 180
          pattern: ^[-a-zA-Z0-9_]+$
        parent:
          type: string
          format: uuid
          nullable: true
        image_url:
          nullable: true
          oneOf:
          - type: string
            format: uri
          - type: string
            maxLength: 0
        sort_order:
          type: integer
          minimum: 0
          default: 0
        is_active:
          type: boolean
      required:
      - name
    CategoryListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTree'
        meta:
          $ref: '#/components/schemas/CatalogPaginationMeta'
      required:
      - data
      - message
      - meta
      - success
    CategoryReorder:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryReorderItem'
      required:
      - items
    CategoryReorderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        sort_order:
          type: integer
          minimum: 0
      required:
      - id
      - sort_order
    CategoryReorderResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTree'
      required:
      - data
      - message
      - success
    CategoryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/CategoryTree'
      required:
      - data
      - message
      - success
...
    CategoryTree:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        parent_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        name:
          type: string
          readOnly: true
        slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        image_url:
          type: string
          readOnly: true
          nullable: true
        sort_order:
          type: integer
          readOnly: true
        is_active:
          type: boolean
          readOnly: true
        children:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - children
      - created_at
      - id
      - image_url
      - is_active
      - name
      - parent_id
      - slug
      - sort_order
      - updated_at
    CategoryTreeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTree'
      required:
      - data
      - message
      - success
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: {}
      required:
      - data
      - message
      - success
...
    PatchedBrandUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        slug:
          type: string
          maxLength: 180
          pattern: ^[-a-zA-Z0-9_]+$
        logo_url:
          nullable: true
          oneOf:
          - type: string
            format: uri
          - type: string
            maxLength: 0
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
    PatchedCategoryUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        slug:
          type: string
          maxLength: 180
          pattern: ^[-a-zA-Z0-9_]+$
        parent:
          type: string
          format: uuid
          nullable: true
        image_url:
          nullable: true
          oneOf:
          - type: string
            format: uri
          - type: string
            maxLength: 0
        sort_order:
          type: integer
          minimum: 0
        is_active:
          type: boolean
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT